combat: hit entry uses the hurt rect, not the collide box (monhun-ardu-q0o)
Owner bug: a sword visibly overlapping the chicken body dealt no damage.
Game::target.rect (the melee/whirl entry gate) was set to the authored collide
box -- the chicken's 9x9 legs, the bull's hooves -- so the drawn body was
undamageable. target.rect is now monsterHurtRect: the body box unioned with the
present head/appendage zones in the current facing (same mirror the hit resolve
uses). Collision keeps monsterCollideRect via pushApart, unchanged.

Tests: bull/heavy/world pins updated off the collide rect; T1 chicken
body-only, T2 heavy tail-only east, T3 tail west mirror, T4 out-of-reach
negative. Shipping 29316/29696 (380 free); dev-hitboxes 29408/29696 (288 free)
after the wireBox int16 trim.

+++ src/render.hpp
@@ -1081,7 +1081,7 @@ static void drawEffects(const mh::Game &g, int16_t camX, int16_t camY) {
 
 // Single-pixel open-square border: solid for hurt boxes, dotted (every other
 // pixel) for hit boxes. Border only -- the interior is never touched.
-static void wireBox(int32_t x, int32_t y, int32_t w, int32_t h, bool dotted) {
+static void wireBox(int16_t x, int16_t y, int16_t w, int16_t h, bool dotted) {
     if (w < 1 || h < 1)
         return;
     if (!dotted) {
@@ -1091,11 +1091,11 @@ static void wireBox(int32_t x, int32_t y, int32_t w, int32_t h, bool dotted) {
         blk(x + w - 1, y, 1, h, 3);
         return;
     }
-    for (int32_t i = 0; i < w; i += 2) {
+    for (int16_t i = 0; i < w; i += 2) {
         blk(x + i, y, 1, 1, 3);
         blk(x + i, y + h - 1, 1, 1, 3);
     }
-    for (int32_t j = 0; j < h; j += 2) {
+    for (int16_t j = 0; j < h; j += 2) {
         blk(x, y + j, 1, 1, 3);
         blk(x + w - 1, y + j, 1, 1, 3);
     }
